Silence some warnings

+++ gfx/drivers/gl3.c
@@ -685,7 +685,7 @@ static void gfx_display_gl3_draw(gfx_display_ctx_draw_t *draw,
    if (gl->chain.active)
    {
       glActiveTexture(GL_TEXTURE0);
-      glBindTexture(GL_TEXTURE_2D, draw->texture);
+      glBindTexture(GL_TEXTURE_2D, (GLuint)draw->texture);
 
       gl->chain.shader->set_coords(gl->chain.shader_data, draw->coords);
       gl->chain.shader->set_mvp(gl->chain.shader_data,
@@ -703,7 +703,7 @@ static void gfx_display_gl3_draw(gfx_display_ctx_draw_t *draw,
          *loc                   = NULL;
 
       glActiveTexture(GL_TEXTURE1);
-      glBindTexture(GL_TEXTURE_2D, draw->texture);
+      glBindTexture(GL_TEXTURE_2D, (GLuint)draw->texture);
 
       switch (draw->pipeline_id)
       {
